Accept --color=always and --color=never options.
For the --color option, accept "always" as an alias for "yes" and
"never" as an alias for "no", because GNU diff uses "always" and "never"
instead of "yes" and "no".  This allows colordiff to operate as expected
if used as a drop-in replacement for a GNU diff command line already
containing "--color=always", which is better than forcing the user to
change the option to "--color=yes".